Update (/rotorcraft) aircraft.json with new eVTOLs
Some eVTOL models are proposed:

- Joby S4
- VoloCity
- Ehang 216-s
- EVE
- Cora (from Wisk Aero)
- Lilium Jet
- Midnight (from Archer)
I did not include Tecnalia (from UMILES): It is a defunct aircraft that is becoming "Crisalion Mobility" but still a prototype. 

I proposed the models that are highly expected to be commercialised (are or are going to be certified part 135). These specs have been double-checked (sources: respective websites + evtol.news). However, they need to be accurately modified by the community when more specs become available/accessible to the public.